+++ src/data_loader/optimize.jl
@@ -84,7 +84,7 @@ _copy(qp::QPT) = (q = copy(qp.q), p = copy(qp.p))
 _copy(t::Tuple{<:QPTOAT, <:QPTOAT}) = _copy.(t)
 
 function (o::Optimizer)(nn::NeuralNetwork,
-        dl::DataLoader,
+        dl::Union{DataLoader, ParametricDataLoader},
         batch::Batch,
         n_epochs::Integer,
         loss::NetworkLoss,